refactor(ProductDetail): rename component
- 렌더링 관심사 분리를 위해 컨테이너로 이름 변경
